requirements.txt
flask
pillow
safetensors
torch
torchvision